+++ crates/modular-decomposition/CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 + `MDTree::nodes(module)` function
 
+### Changed
+
++ Docs
+
 ## [0.2.0] - 2024-06-10
 
 ### Added