+++ crates/modular-decomposition/CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Changed
 
++ CI: Replaced actions-rs/toolchain with dtolnay/rust-toolchain
 + CI: Updated actions/checkout
 + Docs