+++ crates/modular-decomposition/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [0.3.0] - 2024-06-25
+
 ### Added
 
 + Examples to README
@@ -56,7 +58,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 + Initial release
 
-[unreleased]: https://github.com/jonasspinner/modular-decomposition/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/jonasspinner/modular-decomposition/compare/v0.3.0...HEAD
+
+[0.3.0]: https://github.com/jonasspinner/modular-decomposition/compare/v0.2.0...v0.3.0
 
 [0.2.0]: https://github.com/jonasspinner/modular-decomposition/compare/v0.1.0...v0.2.0